Refactor building of ODataQueryOptions object
